index.jsにログイン機能追加

+++ index.js
@@ -2,7 +2,7 @@ const express = require("express")
 const app = express()
 const cors = require("cors")
 const jwt = require("jsonwebtoken")
-//const auth = require("./utils/auth")
+const auth = require("./utils/auth")
 
 app.use(cors())
 app.use(express.urlencoded({extended:true}))
@@ -132,23 +132,33 @@ app.get("/item/:id",async(req,res)=>{
 })
 
 //Update Item
-app.put("/item/update/:id",async(req,res)=>{
+app.put("/item/update/:id",auth,async(req,res)=>{
     console.log(req)
     try{
         await connectDB()
-        await ItemModel.updateOne({_id:req.params.id},req.body)
-        return res.status(200).json({message:"アイテム読み取り成功（シングル）",singleItem:singleItem})
+        const singleItem = await ItemModel.findById(req.params.id)
+        if(singleItem.email === req.body.email){
+            await ItemModel.updateOne({_id:req.params.id},req.body)
+            return res.status(200).json({message:"アイテム読み取り成功（シングル）",singleItem:singleItem})
+        }else{
+            throw new Error()
+        }
     }catch(err){
-        return res.status(400).json({message:"アイテム編集失敗"})
+        return res.status(400).json({message:"アイテム編集"})
     }
 })
 
 //Delete Item
-app.delete("/item/delete/:id",async(req,res)=>{
+app.delete("/item/delete/:id",auth,async(req,res)=>{
     try{
         await connectDB()
-        await ItemModel.deleteOne({_id:req.params.id})
-        return res.status(200).json({message:"アイテム削除成功"})
+        const singleItem = await ItemModel.findById(req.params.id)
+        if(singleItem.email === req.body.email){
+            await ItemModel.deleteOne({_id:req.params.id})
+            return res.status(200).json({message:"アイテム削除成功"})
+        }else{
+            throw new Error()
+        }
     }catch(err){
         return res.status(400).json({message:"アイテム削除失敗"})
     }
@@ -179,7 +189,6 @@ app.post("/user/login",async(req,res)=>{
                     email:req.body.email,
                 }
                 const token = jwt.sign(payload,secret_key,{expiresIn:"23h"})
-                console.log(token)
                 return res.status(200).json({message:"ログイン成功",token:token})
             }else{
                 return res.status(400).json({message:"ログイン失敗パスワードが違います"})
